Add support for multiple companies

+++ app/controllers/admin/UsersController.php
@@ -14,6 +14,7 @@ use DB;
 use Input;
 use User;
 use Asset;
+use Company;
 use Lang;
 use Actionlog;
 use Location;
@@ -49,6 +50,7 @@ class UsersController extends AdminController {
         'email' => 'email|unique:users,email',
         'password' => 'required|min:6',
         'password_confirm' => 'required|min:6|same:password',
+        'company_id' => 'integer',
     );
 
     /**
@@ -84,6 +86,7 @@ class UsersController extends AdminController {
 
         $location_list = locationsList();
         $manager_list = managerList();
+        $company_list = Company::getSelectList();
 
         /* echo '<pre>';
           print_r($userPermissions);
@@ -95,6 +98,7 @@ class UsersController extends AdminController {
         return View::make('backend/users/edit', compact('groups', 'userGroups', 'permissions', 'userPermissions'))
                         ->with('location_list', $location_list)
                         ->with('manager_list', $manager_list)
+                        ->with('company_list', $company_list)
                         ->with('user', new User);
     }
 
@@ -121,6 +125,8 @@ class UsersController extends AdminController {
             // permissions here before we create the user.
             // Get the inputs, with some exceptions
             $inputs = Input::except('csrf_token', 'password_confirm', 'groups', 'email_user');
+
+            $inputs['company_id'] = Company::getIdForUser(Input::get('company_id'));
 
             // @TODO: Figure out WTF I need to do this.
             if ($inputs['manager_id'] == '') {
@@ -254,6 +260,10 @@ class UsersController extends AdminController {
             // Get the user information
             $user = Sentry::getUserProvider()->findById($id);
 
+            if (!Company::isCurrentUserHasAccess($user)) {
+                return Redirect::route('users')->with('error', Lang::get('general.insufficient_permissions'));
+            }
+
             // Get this user groups
             $userGroups = $user->groups()->lists('group_id', 'name');
 
@@ -269,6 +279,7 @@ class UsersController extends AdminController {
             $this->encodeAllPermissions($permissions);
 
             $location_list = array('' => '') + Location::lists('name', 'id');
+            $company_list = Company::getSelectList();
             $manager_list = array('' => 'Select a User') + DB::table('users')
                             ->select(DB::raw('concat(last_name,", ",first_name," (",email,")") as full_name, id'))
                             ->whereNull('deleted_at')
@@ -287,6 +298,7 @@ class UsersController extends AdminController {
         // Show the page
         return View::make('backend/users/edit', compact('user', 'groups', 'userGroups', 'permissions', 'userPermissions'))
                         ->with('location_list', $location_list)
+                        ->with('company_list', $company_list)
                         ->with('manager_list', $manager_list);
     }
 
@@ -311,6 +323,10 @@ class UsersController extends AdminController {
         try {
             // Get the user information
             $user = Sentry::getUserProvider()->findById($id);
+
+            if (!Company::isCurrentUserHasAccess($user)) {
+                return Redirect::route('users')->with('error', Lang::get('general.insufficient_permissions'));
+            }
         } catch (UserNotFoundException $e) {
             // Prepare the error message
             $error = Lang::get('admin/users/message.user_not_found', compact('id'));
@@ -359,6 +375,7 @@ class UsersController extends AdminController {
             $user->jobtitle = Input::get('jobtitle');
             $user->phone = Input::get('phone');
             $user->location_id = Input::get('location_id');
+            $user->company_id = Company::getIdForUser(Input::get('company_id'));
             $user->manager_id = Input::get('manager_id');
             $user->notes = Input::get('notes');
 
@@ -490,7 +507,10 @@ class UsersController extends AdminController {
         } else {
             $statuslabel_list = statusLabelList();
             $user_raw_array = array_keys(Input::get('edit_user'));
-            $users = User::whereIn('id', $user_raw_array)->with('groups')->get();
+
+            $users = User::whereIn('id', $user_raw_array)->with('groups');
+            $users = Company::scopeCompanyables($users)->get();
+
             return View::make('backend/users/confirm-bulk-delete', compact('users', 'statuslabel_list'));
         }
     }
@@ -514,7 +534,9 @@ class UsersController extends AdminController {
 
                 $assets = Asset::whereIn('assigned_to', $user_raw_array)->get();
                 $accessories = DB::table('accessories_users')->whereIn('assigned_to', $user_raw_array)->get();
-                $users = User::whereIn('id', $user_raw_array)->delete();
+
+                $users = User::whereIn('id', $user_raw_array);
+                $users = Company::scopeCompanyables($users)->delete();
 
                 foreach ($assets as $asset) {
 
@@ -574,14 +596,20 @@ class UsersController extends AdminController {
             // Get user information
             $user = Sentry::getUserProvider()->createModel()->withTrashed()->find($id);
 
-            // Restore the user
-            $user->restore();
+            if (!Company::isCurrentUserHasAccess($user)) {
+                return Redirect::route('users')->with('error', Lang::get('general.insufficient_permissions'));
+            }
+            else
+            {
+                // Restore the user
+                $user->restore();
 
-            // Prepare the success message
-            $success = Lang::get('admin/users/message.success.restored');
+                // Prepare the success message
+                $success = Lang::get('admin/users/message.success.restored');
 
-            // Redirect to the user management page
-            return Redirect::route('users')->with('success', $success);
+                // Redirect to the user management page
+                return Redirect::route('users')->with('success', $success);
+            }
         } catch (UserNotFoundException $e) {
             // Prepare the error message
             $error = Lang::get('admin/users/message.user_not_found', compact('id'));
@@ -604,7 +632,12 @@ class UsersController extends AdminController {
         $userlog = $user->userlog->load('assetlog', 'consumablelog', 'assetlog.model', 'licenselog', 'accessorylog', 'userlog', 'adminlog');
 
         if (isset($user->id)) {
-            return View::make('backend/users/view', compact('user', 'userlog'));
+
+            if (!Company::isCurrentUserHasAccess($user)) {
+                return Redirect::route('users')->with('error', Lang::get('general.insufficient_permissions'));
+            } else {
+                return View::make('backend/users/view', compact('user', 'userlog'));
+            }
         } else {
             // Prepare the error message
             $error = Lang::get('admin/users/message.user_not_found', compact('id'));
@@ -691,6 +724,7 @@ class UsersController extends AdminController {
             $this->encodeAllPermissions($permissions);
 
             $location_list = array('' => '') + Location::lists('name', 'id');
+            $company_list = Company::getSelectList();
             $manager_list = array('' => 'Select a User') + DB::table('users')
                             ->select(DB::raw('concat(last_name,", ",first_name," (",email,")") as full_name, id'))
                             ->whereNull('deleted_at')
@@ -702,6 +736,7 @@ class UsersController extends AdminController {
             // Show the page
             return View::make('backend/users/edit', compact('groups', 'userGroups', 'permissions', 'userPermissions'))
                             ->with('location_list', $location_list)
+                            ->with('company_list', $company_list)
                             ->with('manager_list', $manager_list)
                             ->with('user', $user)
                             ->with('clone_user', $user_to_clone);
@@ -783,6 +818,7 @@ class UsersController extends AdminController {
                             'password' => $pass,
                             'activated' => $activated,
                             'location_id' => $row[4],
+                            'company_id' => Company::getIdForUser($row[5]),
                             'permissions' => '{"user":1}',
                             'notes' => 'Imported user'
                         );
@@ -847,7 +883,9 @@ class UsersController extends AdminController {
             $sort = e(Input::get('sort'));
         }
 
-        $users = User::select(array('users.id','users.email','users.username','users.location_id','users.manager_id','users.first_name','users.last_name','users.created_at','users.notes'))->with('assets', 'accessories', 'consumables', 'licenses', 'manager', 'sentryThrottle', 'groups', 'userloc');
+        $users = User::select(array('id','email','username','location_id','manager_id','first_name','last_name','created_at','notes','company_id'))
+            ->with('assets','accessories','consumables','licenses','manager','sentryThrottle','groups','userloc','company');
+        $users = Company::scopeCompanyables($users);
 
         switch ($status) {
         case 'deleted':
@@ -929,7 +967,8 @@ class UsersController extends AdminController {
                 'consumables'        => $user->consumables->count(),
                 'groups'        => $group_names,
                 'notes'         => $user->notes,
-                'actions'       => ($actions) ? $actions : ''
+                'actions'       => ($actions) ? $actions : '',
+                'companyName'   => is_null($user->company) ? '' : e($user->company->name)
             );
         }
 
@@ -951,7 +990,10 @@ class UsersController extends AdminController {
 
         if (isset($user->id)) {
 
-            if (Input::hasFile('userfile')) {
+            if (!Company::isCurrentUserHasAccess($user)) {
+                return Redirect::route('users')->with('error', Lang::get('general.insufficient_permissions'));
+            }
+            else if (Input::hasFile('userfile')) {
 
                 foreach (Input::file('userfile') as $file) {
 
@@ -1012,13 +1054,19 @@ class UsersController extends AdminController {
         // the license is valid
         if (isset($user->id)) {
 
-            $log = Actionlog::find($fileId);
-            $full_filename = $destinationPath . '/' . $log->filename;
-            if (file_exists($full_filename)) {
-                unlink($destinationPath . '/' . $log->filename);
+            if (!Company::isCurrentUserHasAccess($user)) {
+                return Redirect::route('users')->with('error', Lang::get('general.insufficient_permissions'));
             }
-            $log->delete();
-            return Redirect::back()->with('success', Lang::get('admin/users/message.deletefile.success'));
+            else
+            {
+                $log = Actionlog::find($fileId);
+                $full_filename = $destinationPath . '/' . $log->filename;
+                if (file_exists($full_filename)) {
+                    unlink($destinationPath . '/' . $log->filename);
+                }
+                $log->delete();
+                return Redirect::back()->with('success', Lang::get('admin/users/message.deletefile.success'));
+            }
         } else {
             // Prepare the error message
             $error = Lang::get('admin/users/message.does_not_exist', compact('id'));
@@ -1040,9 +1088,15 @@ class UsersController extends AdminController {
 
         // the license is valid
         if (isset($user->id)) {
-            $log = Actionlog::find($fileId);
-            $file = $log->get_src();
-            return Response::download($file);
+            if (!Company::isCurrentUserHasAccess($user)) {
+                return Redirect::route('users')->with('error', Lang::get('general.insufficient_permissions'));
+            }
+            else
+            {
+                $log = Actionlog::find($fileId);
+                $file = $log->get_src();
+                return Response::download($file);
+            }
         } else {
             // Prepare the error message
             $error = Lang::get('admin/users/message.does_not_exist', compact('id'));
